+++ OneBlink.SDK/models/FormApprovalPrefillPathSegment.cs
@@ -1,0 +1,16 @@
+using System;
+
+namespace OneBlink.SDK.Model
+{
+    public class FormApprovalPrefillPathSegment
+    {
+        public string kind
+        {
+            get; set;
+        }
+        public Guid formElementId
+        {
+            get; set;
+        }
+    }
+}